Create 4_FundMe.sol
Finally, a contract which deals with crypto currency (Ethereum).
This have functions which deal with Meta Mask Wallet of SepliaETH test net. This have functions such as Withdraw and Fund, along with a track of record that have Funders in it. Go through the link in the ReadMe file to learn more.