Create 5_PriceConverter.sol
As Blockchain wants to disconnects itself from the server based network, it should be relied on the data given by a server based system. If it does there is no meaning for doing this. So block chains use a special type of networks to get Data feeds.
Go through "docs.chain.link" to get more details.